+++ x/group/keeper/msg_server_test.go
@@ -405,7 +405,8 @@ func (s *TestSuite) TestUpdateGroupMembers() {
 				{
 					Member:  &group.Member{Address: member2, Weight: "2", AddedAt: s.blockTime},
 					GroupId: groupID,
-				}},
+				},
+			},
 		},
 		"remove existing member": {
 			req: &group.MsgUpdateGroupMembers{
@@ -424,7 +425,8 @@ func (s *TestSuite) TestUpdateGroupMembers() {
 				{
 					Member:  &group.Member{Address: member2, Weight: "2", AddedAt: s.blockTime},
 					GroupId: groupID,
-				}},
+				},
+			},
 		},
 		"remove unknown member": {
 			req: &group.MsgUpdateGroupMembers{
@@ -447,7 +449,8 @@ func (s *TestSuite) TestUpdateGroupMembers() {
 				}, {
 					Member:  &group.Member{Address: member2, Weight: "2", AddedAt: s.blockTime},
 					GroupId: groupID,
-				}},
+				},
+			},
 		},
 		"with wrong admin": {
 			req: &group.MsgUpdateGroupMembers{